Remove CPAPMonitor class and CS_SENSE functionality; add diagnostics API endpoint
CPAPMonitor class deleted:
- include/CPAPMonitor.h removed (full class + stub)
- src/CPAPMonitor.cpp removed (all monitoring logic)
- CpapWebServer constructor no longer takes cpapMonitor parameter
- All CPAPMonitor references removed from CpapWebServer.h/cpp

Reason: CS_SENSE pin detection unreliable in hardware, feature disabled by default and unused.

New /api/diagnostics endpoint:
- Returns free_heap and max_all

+++ src/TrafficMonitor.cpp
@@ -156,8 +156,8 @@ uint32_t TrafficMonitor::getTotalIdleSamples() const {
     return _totalIdleSamples;
 }
 
-uint16_t TrafficMonitor::getLastPulseCount() const {
-    return _lastPulseCount;
+uint32_t TrafficMonitor::getLastPulseCount() const {
+    return (uint32_t)_lastPulseCount;
 }
 
 void TrafficMonitor::resetStatistics() {
